fix bot claims
previously the accept claim or reject claim actions could be overwritten
and forgotten if another event came in after the claim event but before
the accept or reject action was handled.

instead, accepting or rejecting claims should not be actions; claims
should just be checked for and dealt with on every handling loop.